+++ projects/ngx-scroll-animations/src/lib/ngx-scroll-animations.directive.ts
@@ -2,6 +2,7 @@ import {
   afterNextRender,
   AfterViewInit,
   booleanAttribute,
+  computed,
   DestroyRef,
   Directive,
   ElementRef,
@@ -34,7 +35,7 @@ import {
   IS_BROWSER,
   providePlatformDetection,
 } from './provider/platform.provider';
-import { ThresholdModeT } from './utils/ngx-scroll-animations-types';
+import { ThresholdModeT } from './types/ngx-scroll-animations.types';
 
 @Directive({
   standalone: true,
@@ -61,18 +62,54 @@ export class NgxScrollAnimationsDirective
   public endAnimation = output();
 
   /**
-   * The speed at which the animation runs in milliseconds.
-   * @param speed - The desired speed of the animation, chosen from predefined options.
+   * @deprecated Use `duration` instead.
+   * TODO: Remove deprecated
+   */
+  public speed = input<string | undefined, string | number | undefined>(
+    undefined,
+    {
+      transform: (value: string | number | undefined) => {
+        if (value === undefined) return value;
+        return `${numberAttribute(value, 300)}ms`;
+      },
+    }
+  );
+
+  /**
+   * The duration at which the animation runs in milliseconds.
    * @default 300
    */
-  public speed = input(`${this.configs?.speed ?? 300}ms`, {
+  public duration = input(
+    this.configs?.speed ?? this.configs?.duration ?? 300,
+    {
+      transform: (value) => numberAttribute(value, 300),
+    }
+  );
+
+  /**
+   * The duration (in milliseconds) for playing the animation in reverse when the element leaves the viewport.
+   * This applies only when the animation runs multiple times.
+   *
+   * @default duration / 2
+   */
+  public reverseDuration = input<
+    number | undefined,
+    string | number | undefined
+  >(this.configs?.reverseDuration, {
     transform: (value) => {
-      return `${numberAttribute(value, 300)}ms`;
+      return numberAttribute(value, 300);
     },
   });
+
+  private _reverseDuration = computed(() => {
+    if (this.reverseDuration() !== undefined) {
+      return this.reverseDuration();
+    }
+    return this.duration() / 2;
+  });
+
   /**
    * Delays the start of the animation. Accepts the delay time in milliseconds.
-   * @param delayTime - The time in milliseconds to delay the start of the animation.
    * @default 0
    */
   public delay = input(this.configs?.delay?.toString() ?? '', {
@@ -89,13 +126,21 @@ export class NgxScrollAnimationsDirective
   });
 
   /**
+   * @deprecated Use `timingFunction` instead.
+   * TODO: Remove deprecated
+   */
+  public easing = input<string | undefined>(undefined);
+
+  /**
    * Defines how the animation accelerates and decelerates during its runtime.
    *
    * Example values: `ease`, `ease-in`, `ease-out`, `ease-in-out`, `cubic-bezier()`
    *
    * @default `ease`
    */
-  public easing = input<string>(this.configs?.easing ?? 'ease');
+  public timingFunction = input<string>(
+    this.configs?.timingFunction ?? this.configs?.easing ?? 'ease'
+  );
 
   /**
    * Defines the mode for calculating the threshold: `percent` or `pixel`.
@@ -138,15 +183,28 @@ export class NgxScrollAnimationsDirective
   });
 
   /**
-   * The gap between the animation start point and the animation end point.
+   * @deprecated Use `triggerBuffer` instead.
+   */
+  public undoGap = input<number | undefined, string | number | undefined>(
+    undefined,
+    {
+      transform: (value) => numberAttribute(value, 20),
+    }
+  );
+
+  /**
+   * The buffer between the animation start point and the animation end point.
    * This prevents the animation from starting/stopping improperly,
    * which often happens on mobile devices.
    *
    * @default 20
    */
-  public undoGap = input(this.configs?.undoGap ?? 20, {
-    transform: (value) => numberAttribute(value, 20),
-  });
+  public triggerBuffer = input(
+    this.configs?.undoGap ?? this.configs?.triggerBuffer ?? 20,
+    {
+      transform: (value) => numberAttribute(value, 20),
+    }
+  );
 
   /**
    * Specifies the animation to be played.
@@ -170,6 +228,7 @@ export class NgxScrollAnimationsDirective
   public lastViewState?: VisibleState;
 
   private temporaryBounding!: DOMRect;
+  private temporaryY = 0;
 
   private onStable = new Subject<void>();
 
@@ -266,6 +325,7 @@ export class NgxScrollAnimationsDirective
           }
           return undefined;
         }),
+        distinctUntilChanged(),
         // Destroy the animation listener
         takeUntilDestroyed(this.destroyRef)
       )
@@ -296,9 +356,15 @@ export class NgxScrollAnimationsDirective
       this.count++;
       this.clearAnimation(true);
       this.triggerIdle();
-      this.elRef.nativeElement.style.animation = `${this.animationType()} ${this.speed()} ${
-        this.delay() ? this.delay() + ' ' : ''
-      }${this.easing()} ${direction}`;
+      const duration =
+        direction === 'normal' ? this.duration() : this._reverseDuration();
+
+      const delay = this.delay() ? this.delay() + ' ' : '';
+      const timingFunction = this.easing() ?? this.timingFunction();
+
+      this.elRef.nativeElement.style.animation = `${this.animationType()} ${
+        this.speed() ?? `${duration}ms`
+      } ${delay}${timingFunction} ${direction}`;
     }
   }
 
@@ -327,10 +393,6 @@ export class NgxScrollAnimationsDirective
 
   /**
    * Calculates the visibility of the element within the viewport, based on the given threshold.
-   * @param element - The ElementRef of the HTML element to check.
-   * @param threshold - The threshold value for visibility.
-   * @param thresholdMode - The mode ('percent' or 'pixel') for calculating the threshold.
-   * @param undoGap - The gap between the animation start point and animation leave point.
    * @returns An Observable emitting the visibility state (0, 1, or undefined).
    */
   private checkVisibility(): Observable<VisibleState> {
@@ -344,11 +406,14 @@ export class NgxScrollAnimationsDirective
             const realRect = this.elRef.nativeElement.getBoundingClientRect();
 
             let elementHeight = realRect.height;
+            let y = realRect.y;
 
             if (this.activeAnimations > 0) {
               elementHeight = this.temporaryBounding.height;
+              y = this.temporaryY - window.scrollY;
             } else {
               this.temporaryBounding = realRect;
+              this.temporaryY = realRect.y + window.scrollY;
             }
 
             const triggerPos =
@@ -356,10 +421,10 @@ export class NgxScrollAnimationsDirective
                 ? elementHeight * this.threshold()
                 : this.threshold();
 
-            const currentPos = realRect.top - view.height + triggerPos;
+            const currentPos = y - view.height + triggerPos;
             if (currentPos <= 0) {
               return 'visible';
-            } else if (currentPos > this.undoGap()) {
+            } else if (currentPos > (this.undoGap() ?? this.triggerBuffer())) {
               return 'hidden';
             } else {
               return 'hold';